Split based on commas in CC environment variable for chapel_commits emails
We're currently observing that a comma-separated list of email
addresses in the GITHUB_COMMIT_EMAILER_RECIPIENT_CC environment
variable is only sending to the first address rather than all of them,
in spite of the fact that the message header looks as though it
contains all of them using a reasonable format.

This applies a .split(",") to the variable before passing it to the
envelopes initializer to see whether splitting it into a list of
addresses works better.