Remove any attempt at dashed separators in chapel-commits mails, for now
My attempt to change from '----' to '––––' in our separators ran
afoul of some Python 2 vs. Python 3 issues due to its use of
unicode, and wrestling with it is taking more time than I think
it's worth for now, so this PR stops trying to create visual
separators to try and get commit mails working with Discourse.